+++ src/ShadowDrop.Cli/Configuration/EnvironmentReader.cs
@@ -1,0 +1,8 @@
+// Copyright (c) 2026 Christian Flessa. All rights reserved.
+// This file is licensed under the MIT license. See LICENSE in the project root for more information.
+namespace ShadowDrop.Cli.Configuration;
+
+internal sealed class EnvironmentReader : IEnvironmentReader
+{
+    public String? GetEnvironmentVariable(String variableName) => Environment.GetEnvironmentVariable(variableName);
+}